Remove old powered by expo
